Puts shared sorting method in its own concern
